requirements.txt
requests==2.32.3
pillow==10.4.0
pyzbar==0.1.9
tqdm==4.66.4
qreader==3.14
opencv-python==4.10.0.84
pandas==2.2.2